Fixed issue where mouse position isn't being detected correctly
changed to divide by block_width/height, no need for render_x/y in highlight function

+++ src/Grid.h
@@ -11,7 +11,7 @@ public:
   float block_width;
   float block_height;
 
-  std::vector<int> pixels;
+  std::vector<float> pixels;
 
   Grid(int size_x, int size_y,int screen_width, int screen_height);
   void DrawGrid(SDL_Renderer *renderer);
